requirements.txt
streamlit
google-generativeai
psycopg2-binary
pandas
Pillow
python-dotenv
streamlit-aggrid
fpdf
python-pptx
python-docx
